+++ CHANGES.md
@@ -17,6 +17,7 @@ Changes in 10.9
 - Remove period from URL when switching School / Year / MP / CP in Side.php
 - HTML Rename "period" select input ID to "school_period" to avoid conflicts in Attendance/, Eligibility/, Food_Service/ & Grades/TeacherCompletion.php
 - CSS Add modname class, ie .modname-grades-reportcards-php for modname=Grades/ReportCards.php in PDF.php
+- CSS Add .list-column-[column_name] class in ListOutput.fnc.php
 
 Changes in 10.8.4
 -----------------